standalone tests for sWeights
this was for testing for the cause of large residuals in phiR and theta
for sWeighted fits; the cause was the MC sample having a lot of
background. The solution was to "trim" the MC tree to remove as much
of the background as possible, so that we are using a MC data set
which is (hopefully) much closer to what the sWeighted data are.